requirements.txt
prefect>=2.0.0
pandas>=1.5.0
numpy>=1.20.0
sqlalchemy>=1.4.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=0.19.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0 
pycoingecko>=2.0.0
pyarrow>=14.0.1
polars-lts-cpu>=0.20.0
supabase>=2.0.0
pytest-mock>=3.10.0
dune-spice>=0.2.5
networkx>=3.0.0
scipy>=1.11.0
anthropic==0.52.0
extruct==0.18.0
